Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/IndexedDB/fire-upgradeneeded-event-exception-expected.txt

CONSOLE MESSAGE: Error
CONSOLE MESSAGE: Error
CONSOLE MESSAGE: Error
CONSOLE MESSAGE: TypeError: 'handleEvent' property of event listener should be callable
CONSOLE MESSAGE: Error
CONSOLE MESSAGE: Error

PASS Exception in upgradeneeded handler
PASS Exception in upgradeneeded listener
PASS Exception in upgradeneeded "handleEvent" lookup
PASS Exception in upgradeneeded due to non-callable "handleEvent"
PASS Exception in second upgradeneeded listener
PASS Exception in first upgradeneeded listener, tx active in second
